Moving code that draws the log levels to its own class
this allows other scripts to use the same code to draw the log levels
