+++ Assets/Mirror/Editor/Logging/LogLevelWindow.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
 
@@ -16,21 +15,9 @@ namespace Mirror.EditorScripts.Logging
             EditorGUILayout.EndVertical();
 
             EditorGUILayout.BeginVertical(EditorStyles.inspectorDefaultMargins);
-            foreach (KeyValuePair<string, ILogger> item in LogFactory.loggers)
-            {
-                DrawLoggerField(item);
-            }
+            LogLevelsGUI.DrawLogFactoryDictionary();
             EditorGUILayout.EndVertical();
         }
-
-        static void DrawLoggerField(KeyValuePair<string, ILogger> item)
-        {
-            ILogger logger = item.Value;
-            string name = item.Key;
-
-            logger.filterLogType = (LogType)EditorGUILayout.EnumPopup(new GUIContent(name), logger.filterLogType);
-        }
-
 
         [MenuItem("Window/Analysis/Mirror Log Levels", priority = 20002)]
         public static void ShowWindow()
